feat: Implement 'if' statements and control flow
This commit adds support for conditional execution using `if-else` statements, a fundamental feature for the language.

- **Lexer & AST:** Added `IF` and `END` keywords. Introduced `IfStmt` and `BlockStmt` nodes to the AST to represent the new control flow structures.

- **Parser:** Extended the parser to recognize and parse `if`, `else if`, and `else` blocks, terminated by an `end` keyword.

- **Compiler & VM:** Implemented jump opcodes (`OP_JUMP`, `OP_JUMP_IF_FALSE`) to handle conditional branching. The compiler now uses backpatching to correctly resolve forward jumps. The VM's execution loop has been updated to handle these new opcodes.

- **Testing:** Added a comprehensive suite of tests to verify the correct behavior of various `if-else` constructs, ensuring that the correct branches are executed based on the condition.

+++ tests/test_vm.cpp
@@ -8,7 +8,7 @@ TEST_CASE("testing the VM with OP_CONSTANT") {
 
     int constant = chunk.addConstant(1.2);
     chunk.write(static_cast<uint8_t>(nota::OpCode::OP_CONSTANT), 1);
-    chunk.write(constant, 1);
+    chunk.write(static_cast<uint8_t>(constant), 1);
 
     vm.interpret(&chunk);
 
@@ -21,7 +21,7 @@ TEST_CASE("testing the VM with OP_NEGATE") {
 
     int constant = chunk.addConstant(1.2);
     chunk.write(static_cast<uint8_t>(nota::OpCode::OP_CONSTANT), 1);
-    chunk.write(constant, 1);
+    chunk.write(static_cast<uint8_t>(constant), 1);
     chunk.write(static_cast<uint8_t>(nota::OpCode::OP_NEGATE), 1);
 
     vm.interpret(&chunk);
